pyproject.toml
[build-system]
requires = ["setuptools >= 61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "DNAHash"
description = "A specialized datastructure for counting short DNA sequences for use in Bioinformatics."
version = "0.0.1"
requires-python = ">= 3.10"
dependencies = [
  "pybloomer>=0.0.1",
  "numpy>=1.19.5",
  "nptyping>=2.5.0",
]
authors = [
  {name = "Andrew DalPino", email = "support@andrewdalpino.com"},
]
readme = "README.md"
license = {text = "MIT"}

[project.optional-dependencies]
dev = ["mypy", "biopython", "matplotlib", "PyQt6"]
test = ["mypy"]

[project.urls]
Homepage = "https://github.com/andrewdalpino/DNAHash"
Documentation = "https://github.com/andrewdalpino/DNAHash/README.md"
Source = "https://github.com/andrewdalpino/DNAHash"
